requirements.txt
boto3==1.5.22
slackclient==1.1.2
requests==2.18.4
python_dateutil==2.6.1
